Creation du component client-view
