+++ ErabliereIU/src/app/client-view/agora-call-service/agora-call-service.component.css
@@ -1,0 +1,33 @@
+#video-container {
+    position: fixed;
+    bottom: 20px;
+    right: 20px;
+    max-height: 95%;
+    max-width: 95%;
+    overflow: auto;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+    z-index: 2000;
+    background-color: rgba(129, 129, 129, 0.5);
+    border-radius: 10px;
+    border-width: 2px;
+    border-color: black;
+}
+.player {
+    width: 480px;
+    height: 320px;
+}
+.player-name {
+    margin: 8px 0;
+}
+@media (max-width: 640px) {
+    .player {
+        width: 320px;
+        height: 240px;
+    }
+}
+.select-ellipsis {
+    text-overflow: ellipsis;
+    white-space: nowrap;
+    overflow: hidden;
+    max-width: 300px;
+}
