+++ ErabliereIU/src/app/client-view/client-nav-bar/client-nav-bar.component.html
@@ -1,0 +1,75 @@
+<nav class="navbar navbar-expand-lg navbar-light bd-navbar">
+  <div class="container-fluid">
+    <h2 class="ms-4 me-5">Érablière IU</h2>
+    <button class="navbar-toggler"
+            type="button"
+            data-bs-toggle="collapse"
+            data-bs-target=".navbar-collapse"
+            aria-controls="navbarSupportedContent"
+            aria-expanded="false" aria-label="Toggle navigation">
+      <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="navbar-collapse collapse d-lg-inline-flex flex-lg-row">
+      <ul class="navbar-nav me-auto mb-2 mb-lg-0 me-auto">
+        <li class="nav-item">
+          @if (idErabliereSelectionnee) {
+          <a
+            class="nav-link"
+            routerLink="/e/{{idErabliereSelectionnee}}/graphiques"
+            routerLinkActive="active"
+            ariaCurrentWhenActive="page">
+            Graphiques
+          </a>
+          } @else {
+          <a
+              class="nav-link"
+              routerLink="e"
+              routerLinkActive="active"
+              ariaCurrentWhenActive="page">
+              Graphiques
+          </a>
+          }
+        </li>
+        @for (url of urls; track url) {
+        <li class="nav-item">
+          <a class="nav-link" href="{{ url.href }}" role="button" target="_blank" rel="noopener noreferrer">{{ url.text }}</a>
+        </li>
+        }
+        @if (isLoggedIn && thereIsAtLeastOneErabliere) {
+        <li class="nav-item">
+          <a id="nav-menu-alerte-button" class="nav-link" routerLink="e/{{idErabliereSelectionnee}}/alertes" routerLinkActive="active" ariaCurrentWhenActive="page">
+            Alertes
+          </a>
+        </li>
+        <li class="nav-item">
+          <a id="nav-menu-notes-button" class="nav-link" routerLink="e/{{idErabliereSelectionnee}}/notes" routerLinkActive="active" ariaCurrentWhenActive="page">
+            Notes
+          </a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" routerLink="e/{{idErabliereSelectionnee}}/documentations" routerLinkActive="active" ariaCurrentWhenActive="page">
+            Documentation
+          </a>
+        </li>
+        }
+        <li class="nav-item">
+          <a class="nav-link" routerLink="/apropos" routerLinkActive="active" ariaCurrentWhenActive="page">
+            À Propos
+          </a>
+        </li>
+      </ul>
+      @if (useAuthentication) {
+      <div class="d-flex gap-2">
+        @if (callFeatureEnable && callFeatureEnableForUser) {
+        <agora-call-service></agora-call-service>
+        }
+        @if (isLoggedIn) {
+        <button id="logout-button" class="btn btn-outline-success" (click)="logout()">Déconnexion</button>
+        } @else {
+        <button id="login-button" class="btn btn-outline-success" (click)="login()">Se connecter</button>
+        }
+      </div>
+      }
+    </div>
+  </div>
+</nav>
