+++ ErabliereIU/src/app/client-view/erablieres-side-bar/erablieres-side-bar.component.html
@@ -1,32 +1,45 @@
 <ajouter-erabliere (shouldReloadErablieres)="loadErablieresPage()" />
-<modifier-erabliere 
-    (shouldReloadErablieres)="loadErablieresPage()" 
+<modifier-erabliere
+    (shouldReloadErablieres)="loadErablieresPage()"
     [authEnabled]="!authDisabled"
     [idErabliere]="erabliereSelectionnee?.id" />
 
 <div class="border-right">
     <div class="list-group">
-        <a *ngIf="etat != 'Chargement des erablieres terminé'" class="list-group-item list-group-item-action" href="#">
+        @if (etat !== 'Chargement des erablieres terminé') {
+        <a class="list-group-item list-group-item-action" href="#">
             <div class="d-flex w-100 justify-content-between">
                 <h5 class="mb-1">{{etat}}</h5>
             </div>
         </a>
-        <a *ngFor="let erabliere of erablieres" 
+        }
+        @for (erabliere of erablieres; track  erabliere.id) {
+        <a
             class="list-group-item list-group-item-action"
             [class.active]="erabliereSelectionnee?.id === erabliere.id"
             style="cursor: pointer;"
-            (click)="handleErabliereLiClick(erabliere.id)">{{erabliere.nom}} 
+            (click)="handleErabliereLiClick(erabliere.id)"
+        >
+            {{erabliere.nom}}
+            @if (!erabliere.isPublic || authDisabled || loggedIn) {
             <button
-                *ngIf="(!erabliere.isPublic || authDisabled || loggedIn)" class="btn btn-info btn-sm float-end"
+                class="btn btn-info btn-sm float-end"
                 (click)="openEditErabliereForm(erabliere)" data-bs-toggle="modal"
                 data-bs-target="#modifierErabliereFormModal">&#9998;
             </button>
+            }
         </a>
-        <button *ngIf="loggedIn" class="list-group-item list-group-item-action" data-bs-toggle="modal"
-            data-bs-target="#erabliereFormModal">
+        }
+        @if (loggedIn) {
+        <button
+            class="list-group-item list-group-item-action"
+            data-bs-toggle="modal"
+            data-bs-target="#erabliereFormModal"
+        >
             <div class="d-flex w-100 justify-content-between">
                 <h5 class="mb-1">Ajouter une erablière</h5>
             </div>
         </button>
+        }
     </div>
 </div>
