--- ErabliereIU/src/dashboard/site-nav-bar.component.html
@@ -1,55 +1,0 @@
-<nav class="navbar navbar-expand-lg navbar-light bd-navbar">
-  <div class="container-fluid">
-    <h2 class="ms-4 me-5">Érablière IU</h2>
-    <button class="navbar-toggler"
-            type="button"
-            data-bs-toggle="collapse"
-            data-bs-target=".navbar-collapse"
-            aria-controls="navbarSupportedContent"
-            aria-expanded="false" aria-label="Toggle navigation">
-      <span class="navbar-toggler-icon"></span>
-    </button>
-    <div class="navbar-collapse collapse d-lg-inline-flex flex-lg-row">
-      <ul class="navbar-nav me-auto mb-2 mb-lg-0 me-auto">
-        <li class="nav-item">
-          <a
-            *ngIf="idErabliereSelectionnee"
-            class="nav-link"
-            routerLink="/e/{{idErabliereSelectionnee}}/graphiques"
-            routerLinkActive="active"
-            ariaCurrentWhenActive="page">
-            Graphique
-          </a>
-          <a
-            *ngIf="!idErabliereSelectionnee"
-            class="nav-link"
-            routerLink="e"
-            routerLinkActive="active"
-            ariaCurrentWhenActive="page">
-            Graphique
-          </a>
-        </li>
-        <li *ngFor="let url of urls" class="nav-item">
-          <a class="nav-link" href="{{ url.href }}" role="button" target="_blank" rel="noopener noreferrer">{{ url.text }}</a>
-        </li>
-        <li *ngIf="isLoggedIn && thereIsAtLeastOneErabliere" class="nav-item">
-          <a id="nav-menu-alerte-button" class="nav-link" routerLink="e/{{idErabliereSelectionnee}}/alertes" routerLinkActive="active" ariaCurrentWhenActive="page">Alerte</a>
-        </li>
-        <li *ngIf="isLoggedIn && thereIsAtLeastOneErabliere" class="nav-item">
-          <a id="nav-menu-notes-button" class="nav-link" routerLink="e/{{idErabliereSelectionnee}}/notes" routerLinkActive="active" ariaCurrentWhenActive="page">Notes</a>
-        </li>
-        <li *ngIf="isLoggedIn && thereIsAtLeastOneErabliere" class="nav-item">
-          <a class="nav-link" routerLink="e/{{idErabliereSelectionnee}}/documentations" routerLinkActive="active" ariaCurrentWhenActive="page">Documentation</a>
-        </li>
-        <li class="nav-item">
-          <a class="nav-link" routerLink="/apropos" routerLinkActive="active" ariaCurrentWhenActive="page">À Propos</a>
-        </li>
-      </ul>
-      <div *ngIf="useAuthentication" class="d-flex gap-2">
-        <agora-call-service *ngIf="callFeatureEnable && callFeatureEnableForUser"></agora-call-service>
-        <button id="login-button" class="btn btn-outline-success" *ngIf="!isLoggedIn" (click)="login()">Se connecter</button>
-        <button id="logout-button" class="btn btn-outline-success" *ngIf="isLoggedIn" (click)="logout()">Déconnexion</button>
-      </div>
-    </div>
-  </div>
-</nav>
